refactor(engine): dedup the backdrop-filter blur into one Backend::apply_backdrop_blur (+ fix Path B clamp)
The backdrop-filter blur pipeline was implemented twice — Path A (layer-tree,
Renderer::handle_backdrop_filter) and Path B (display-list,
Backend::render_backdrop_filter) — six identical steps each (flush → clamp device
rect → copy region → Dual-Kawase blur → composite at the clamped rect). The
off-screen-clamp fix had to be applied twice and then SILENTLY DRIFTED: Path A
clamped with `.round()` before `as u32`, Path B truncated — undersizing Path B
backdrops by up to one device pixel on fractional DPR / sub-pixel CTMs.

Consolidate both onto one inherent method `Backend::apply_backdrop_blur`:
clamp → copy → blur → composite live there once, so the clamp can no longer
drift. Both entry points are `&mut Backend`, so the shared method is the natural
home; each caller keeps only what legitimately differs — sigma extraction, the
bounds→device transform source (layer-walk CTM vs command transform), and child
dispatch (layer recursion vs display-list command dispatch).

The unified clamp uses the rounded form (the documented-correct Path A version),
which FIXES Path B's fractional-DPR undersize. At integer DPR round == truncate,
so output is bit-identical (the readback suite + both paths' backdrop tests stay
green); the only behavior change is Path B fractional-DPR, now matching Path A.

Also removes `RenderContext`'s now-dead `device`/`queue`/`surface_format` fields
(only Path A's inline pipeline read them; the helper sources them from the
offscreen renderer) across the struct + 5 construction sites.

Net −160 lines.

## Verification

- GPU readback suite (`--features enable-wgpu-tests --test-threads 1`): 457
  passed / 0 failed (pure refactor, count unchanged). Path A backdrop tests
  (composite-at-device-rect / honors-translation / clamped-when-offscreen, all
  under DPR) and the Path B test (composites-at-clamped-rect-when-crossing-edge)
  all pass through the unified helper — both paths render identically.
- clippy (default + enable-wgpu-tests) -D warnings, fmt, doc -D warnings: clean.
- ce-kieran review: SHIP (behavior preservation, the deliberate Path B clamp fix,
  the no-offscreen-flush edge unification, borrow soundness at both call sites and
  inside the helper, and the RenderContext field removal all traced clean).
